pyproject.toml
[project]
name = "ihunt"
version = "0.1.5"
description = "Information gathering tool."
authors = [
    { name = "hdks", email = "hdks.bug@gmail.com" }
]
dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "duckduckgo-search>=6.2.9",
    "groq>=0.10.0",
    "requests>=2.32.3",
    "google-generativeai>=0.7.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"ihunt" = "ihunt:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ihunt"]
